Fix featured records error with no ExhibitBuilder
Checks for whether ExhibitBuilder is active were too late to be
effective.

This moves the check to before querying for records, and modifies the
display code to more simply handle when there are no results for a
record type. As a side effect this removes the messages printed when
Exhibit Builder isn't installed, but I don't think they're really
important to maintain.