Fix recent records error if ExhibitBuilder missing
Similar to featured, recent records will try to call an ExhibitBuilder
function as well as query the exhibits table even if the plugin is not
installed.

This change skips the recent display completely if exhibits is the
selected type but the plugin isn't active, skipping both the function
call and the query from total_records.